Moving login form file for consistency
